requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.3
click==6.7
colorlover==0.2.1
dash-core-components==0.2.6
dash-html-components==0.3.6
dash-renderer==0.2.4
dash.ly==0.12.4
decorator==4.0.11
enum34==1.1.6
finsymbols==1.1.1
Flask==0.12
Flask-Cors==3.0.2
functools32==3.2.3.post2
gunicorn==19.6.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.9.5
jsonschema==2.6.0
jupyter-core==4.3.0
MarkupSafe==0.23
nbformat==4.3.0
numpy==1.12.0
pandas==0.19.2
pandas-datareader==0.3.0.post0
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
plotly==2.0.2
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
requests-file==1.4.1
requests-ftp==0.3.1
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.11.15
wheel==0.26.0
